Added math for sword rotation. Only kind of works

+++ include/entity.h
@@ -20,8 +20,8 @@ public:
 	// variables used in Drawing the entity
 	GLuint texture;
 	glm::vec2 Position;
-	int Height;
-	int Width;
+	float Height;
+	float Width;
 	glm::vec2 size;
 	GLfloat rotation;
